++ code to update shopping items
Warning: this code is a little dangerous. We give out the ingredient objects, then take them back and count on them being the same object when we get them back. We then use the update method to update one property before committing the ingredients back to the local store. BUUUUTTTT, we don't do anything to ensure some caller doesn't screw with the object in the mean time. If they do, they could end up editing the recipe without meaning to. I should probably re-design this OR find a way to lock the properties of the object I don't want messed with, but I'm not up for it at the moment. Since I don't implement saving either right now, it's not too important.